pyproject.toml
[tool.poetry]
name = "question_app"
version = "0.3.0"
description = "A comprehensive web application for managing Canvas LMS quiz questions with AI-powered feedback generation, intelligent chat assistant using RAG, comprehensive type safety, and VS Code integration."
authors = ["Bryce Kayanuma <BrycePK@vt.edu>", "Robert Fentress <learn@vt.edu>"]
readme = "README.md"
packages = [{include = "question_app", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
jinja2 = "^3.1.2"
python-dotenv = "^1.0.0"
httpx = "^0.25.2"
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
psycopg2-binary = "^2.9.9"
pgvector = "^0.3.0"
beautifulsoup4 = "^4.12.0"
numpy = "<2.0"
markdown2 = "^2.5.4"
lxml = "^6.0.1"
requests = "^2.31.0"
pyjwt = "^2.8.0"
markdown = "^3.10"
pygments = "^2.19.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
httpx = "^0.25.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
sphinx = "^7.0.0"
sphinx-rtd-theme = "^3.0.2"
myst-parser = "^2.0.0"
mypy = "^1.5.1"
pyright = "^1.1.300"

[tool.poetry.scripts]
start = "question_app.main:start"
dev = "question_app.main:dev"
docs = "scripts.build_docs:main"
docs-simple = "scripts.build_docs_simple:main"
docs-serve = "scripts.serve_docs:main"
lint = "scripts.lint_code:main"
format = "scripts.format_code:main"
type-check = "scripts.type_check:main"
test = "scripts.run_tests:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3